Ability to output to a location
We can now output all content that the site needs to function to a specific location relative to the Python source. Hope to make this relative to the working directory soon. Location to output to is given by the '--output' parameter which the html will be written to. Externals such as css and assets are copied to this location when generating.