Ability to load in template resource data from a single location
This allows to load template data from a single location. Defining the root of a data entry by the name of the file loaded in capitalized. This ensure that no duplicates can exist as we require data to be of a specific format (json).

Also added tests for the new loading functionality.
Also fixed up template variable referencing where needed.